--- src/main/resources/templates/howdy.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-
-<h1>Howdy, <span th:text="${name}">Zenith</span>!</h1>
-<h1 th:text="${favNum}"></h1>
-
-</body>
-</html>
