--- src/main/resources/templates/productList.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-<h1>Create Product</h1>
-  <div th:each="item : ${}" >
-
-  </div>
-
-</body>
-</html>
