--- src/main/resources/templates/products.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-<h1>Create Product</h1>
-<form th:action="@{/products/create}" th:method="POST">
-    <div class="form-group">
-        <label for="name">Name</label>
-        <input id="name" name="name" class="form-control" type="text">
-    </div>
-    <div class="form-group">
-        <label for="price">Price</label>
-        <input id="price" name="price" class="form-control" type="text">
-    </div>
-    <input type="submit" class="btn btn-primary btn-block" value="Submit">
-</form>
-
-</body>
-</html>
