--- src/main/resources/templates/products/create.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head th:replace="partials/header :: head"></head>
-
-<body>
-<nav th:replace="partials/navbar :: navbar"></nav>
-<h1>Create Product</h1>
-
-<a th:href="@{/about?num=7}">About Page</a>
-
-
-<form th:action="@{/products/create}" th:method="POST">
-    <label for="name">Product Name</label>
-    <input id="name" type="text" name="name" placeholder="Enter product name">
-    <label for="price">Price In Cents</label>
-    <input id="price" type="number" name="price" placeholder="Enter price in cents">
-    <button>Create Product</button>
-</form>
-</body>
-</html>
