--- src/main/resources/templates/roll-guess.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-
-    <h1>Roll Guess</h1>
-
-    <h3>You guessed: <span th:text="${guess}"></span> </h3>
-    <h3>This value was rolled: <span th:text="${roll}"></span></h3>
-
-<!--    <h3 th:if="${guess==roll}">Correct</h3>-->
-<!--    <h3 th:if="${guess!=roll}">Incorrect</h3>-->
-    <h3 th:text="${guess==roll ? 'Correct' : 'Incorrect'}"></h3>
-
-
-</body>
-</html>
