--- src/main/resources/templates/rollDice2.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-  <meta charset="UTF-8">
-  <title>Title</title>
-</head>
-<body>
-<!--<nav th:replace="partials :: navbar" />-->
-
-
-<h1>Guess a number!</h1>
-
-  <a href="/roll-dice/1">1</a>
-  <a href="/roll-dice/2">2</a>
-  <a href="/roll-dice/3">3</a>
-  <a href="/roll-dice/4">4</a>
-  <a href="/roll-dice/5">5</a>
-  <a href="/roll-dice/6">6</a>
-
-
-
-</body>
-</html>
